simplified update()
created a private function Game::update_edges() that updates the top end bottom edge of the paddle. this moves that logic out of update and sets things up for collision detection in that i can add logic to calculate the front edge of the paddle as well. i also created an overload that accepts a new y position value for handling edge collisions.