left-right collision detection
basic collision detection with paddles implemented. game is reset if ball passes either paddle. i added several data fields and update_edges() overloads to facilitate collision detection. the paddles now carry a field that denotes left or right as their edge calculations are distinct. to do this update_edges() has an overload for each game-object type as game_ball is unique in that it is the only game_object that must track both left and right edges.